:building_construction: Refactoring & cleanup
