+++ packages/website/src/routes/ReefRoutes/Reef/Charts/index.test.tsx
@@ -37,7 +37,12 @@ test("renders as expected", () => {
     },
   ];
   const { container } = render(
-    <Charts depth={0} dailyData={dailyData} temperatureThreshold={0} />
+    <Charts
+      maxMonthlyMean={null}
+      depth={0}
+      dailyData={dailyData}
+      temperatureThreshold={0}
+    />
   );
   expect(container).toMatchSnapshot();
   console.error = originalError;
