+++ packages/website/src/routes/Surveys/View/Charts.tsx
@@ -1,10 +1,8 @@
 import React from "react";
 import { createStyles, Grid, withStyles, WithStyles } from "@material-ui/core";
-import type { Data } from "../../../../store/Reefs/types";
-import "../../../../common/Chart/backgroundPlugin";
-import "../../../../common/Chart/fillPlugin";
-import "../../../../common/Chart/slicePlugin";
-import ChartWithTooltip from "../../../../common/Chart/ChartWithTooltip";
+import type { Data } from "../../../store/Reefs/types";
+
+import ChartWithTooltip from "../../../common/Chart/ChartWithTooltip";
 
 const Charts = ({ classes, ...rest }: ChartsProps) => {
   return (
